Prime Number Calculator
a small program written with one method to simply state if the number is
prime, and another that when given a range of numbers will list out
every prime number in that range.

issue found with larg arrays it takes a very long time to print the
arrays. Need to find a faster solution.